+++ public/pwa-service-worker.js
@@ -1,0 +1,4 @@
+self.addEventListener('install', function(event) {
+  // Dummy service worker that does nothing,
+  // for mobile browsers "Add to home screen".
+});