+++ releases/translations/pull-translations.sh
@@ -42,6 +42,9 @@ tx pull -f -l sl_SI
 echo "Spanish:"
 tx pull -f -l es
 
+echo "Spanish (Chile):"
+tx pull -f -l es_CL
+
 echo "Spanish (Argentina):"
 tx pull -f -l es_AR