+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and adds the following translations:
 
 and fixes the following bugs:
 
+- [Fix Deleting linked card makes board not load](https://github.com/wekan/wekan/issues/2785).
+  Thanks to marc1006 and xet7.
 - [Fix getStartDayOfWeek once again](https://github.com/wekan/wekan/pull/3061).
   Thanks to marc1006.
 - [Fix shortcuts list and support card shortcuts when hovering